Fix spark package name (#518)
* update spark package
