+++ docs/Documentation/UserGuide/7-Ecosystem Integration/4-Spark IoTDB.md
@@ -49,7 +49,7 @@ mvn clean scala:compile compile install
 ```
 spark-shell --jars spark-iotdb-connector-0.9.0-SNAPSHOT.jar,iotdb-jdbc-0.9.0-SNAPSHOT-jar-with-dependencies.jar
 
-import org.apache.iotdb.sparkdb._
+import org.apache.iotdb.spark.db._
 
 val df = spark.read.format("org.apache.iotdb.sparkdb").option("url","jdbc:iotdb://127.0.0.1:6667/").option("sql","select * from root").load
 
@@ -62,7 +62,7 @@ df.show()
 ```
 spark-shell --jars spark-iotdb-connector-0.9.0-SNAPSHOT.jar,iotdb-jdbc-0.9.0-SNAPSHOT-jar-with-dependencies.jar
 
-import org.apache.iotdb.sparkdb._
+import org.apache.iotdb.spark.db._
 
 val df = spark.read.format("org.apache.iotdb.sparkdb").option("url","jdbc:iotdb://127.0.0.1:6667/").option("sql","select * from root").
                         option("lowerBound", [lower bound of time that you want query(include)]).option("upperBound", [upper bound of time that you want query(include)]).
@@ -129,7 +129,7 @@ You can also use narrow table form which as follows: (You can see part 4 about h
 
 ## from wide to narrow
 ```
-import org.apache.iotdb.sparkdb._
+import org.apache.iotdb.spark.db._
 
 val wide_df = spark.read.format("org.apache.iotdb.sparkdb").option("url", "jdbc:iotdb://127.0.0.1:6667/").option("sql", "select * from root where time < 1100 and time > 1000").load
 val narrow_df = Transformer.toNarrowForm(spark, wide_df)
@@ -137,7 +137,7 @@ val narrow_df = Transformer.toNarrowForm(spark, wide_df)
 
 ## from narrow to wide
 ```
-import org.apache.iotdb.sparkdb._
+import org.apache.iotdb.spark.db._
 
 val wide_df = Transformer.toWideForm(spark, narrow_df)
 ```
@@ -147,7 +147,7 @@ val wide_df = Transformer.toWideForm(spark, narrow_df)
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SparkSession;
-import org.apache.iotdb.sparkdb.*
+import org.apache.iotdb.spark.db.*
 
 public class Example {
