+++ spark-iotdb-connector/src/main/java/org/apache/iotdb/spark/db/SQLConstant.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.sparkdb;
+package org.apache.iotdb.spark.db;
 
 /**
  * this class contains several constants used in SQL.
