+++ spark-iotdb-connector/src/main/scala/org/apache/iotdb/spark/db/Converter.scala
@@ -16,7 +16,7 @@
   * specific language governing permissions and limitations
   * under the License.
   */
-package org.apache.iotdb.sparkdb
+package org.apache.iotdb.spark.db
 
 import java.sql.{Statement, _}
