+++ spark-iotdb-connector/src/test/scala/org/apache/iotdb/spark/db/EnvironmentUtils.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.sparkdb;
+package org.apache.iotdb.spark.db;
 
 import java.io.File;
 import java.io.IOException;
